version  190

+++ app/src/main/java/co/dito/abako/apijitpack/data/model/request/DocumentDayRequest.kt
@@ -31,7 +31,7 @@ data class DetallePedido(
     val idArt: Int,
     val cant: Int,
     val um: String,
-    val descu: Int
+    val descu: Double
 )
 
 data class DocumentDayResponse(
